Move custom properties fallback to preset
