+++ packages/postcss-preset-primer/package.json
@@ -26,6 +26,7 @@
     "babel-jest": "^29.7.0",
     "babel-plugin-transform-import-meta": "^2.2.1",
     "postcss": "^8.4.41",
+    "postcss-custom-properties-fallback": "^1.0.2",
     "postcss-mixins": "^10.0.1",
     "typescript": "^5.5.3"
   }
