+++ packages/react/package.json
@@ -202,7 +202,6 @@
     "mdast-util-to-string": "4.0.0",
     "micromark-extension-frontmatter": "2.0.0",
     "micromark-extension-mdxjs": "3.0.0",
-    "postcss-custom-properties-fallback": "1.0.2",
     "postcss-preset-env": "^10.0.0",
     "postcss-preset-primer": "^0.0.0",
     "react": "18.3.1",
