+++ packages/react/rollup.config.mjs
@@ -1,4 +1,3 @@
-import fs from 'node:fs'
 import path from 'node:path'
 import {fileURLToPath} from 'node:url'
 import commonjs from '@rollup/plugin-commonjs'
@@ -7,11 +6,10 @@ import babel from '@rollup/plugin-babel'
 import replace from '@rollup/plugin-replace'
 import terser from '@rollup/plugin-terser'
 import glob from 'fast-glob'
-import customPropertiesFallback from 'postcss-custom-properties-fallback'
 import {visualizer} from 'rollup-plugin-visualizer'
 import {importCSS} from 'rollup-plugin-import-css'
 import postcss from 'rollup-plugin-postcss'
-import postssPresetPrimer from 'postcss-preset-primer'
+import postcssPresetPrimer from 'postcss-preset-primer'
 import MagicString from 'magic-string'
 import packageJson from './package.json' assert {type: 'json'}
 
@@ -88,46 +86,6 @@ function createPackageRegex(name) {
   return new RegExp(`^${name}(/.*)?`)
 }
 
-const postcssPlugins = [
-  postssPresetPrimer(),
-  customPropertiesFallback({
-    importFrom: [
-      () => {
-        let customProperties = {}
-        const filePaths = glob.sync(['fallbacks/**/*.json', 'docs/functional/themes/light.json'], {
-          cwd: path.join(__dirname, '../../node_modules/@primer/primitives/dist/'),
-          ignore: ['fallbacks/color-fallbacks.json'],
-        })
-
-        for (const filePath of filePaths) {
-          const fileData = fs.readFileSync(
-            path.join(__dirname, '../../node_modules/@primer/primitives/dist/', filePath),
-            'utf8',
-          )
-
-          const jsonData = JSON.parse(fileData)
-          let result = {}
-
-          if (filePath === 'docs/functional/themes/light.json') {
-            for (const variable of Object.keys(jsonData)) {
-              result[`--${variable}`] = jsonData[variable].value
-            }
-          } else {
-            result = jsonData
-          }
-
-          customProperties = {
-            ...customProperties,
-            ...result,
-          }
-        }
-
-        return {customProperties}
-      },
-    ],
-  }),
-]
-
 const postcssModulesOptions = {
   generateScopedName: 'prc-[folder]-[local]-[hash:base64:5]',
 }
@@ -175,7 +133,7 @@ const baseConfig = {
     }),
     importCSS({
       modulesRoot: 'src',
-      postcssPlugins,
+      postcssPlugins: [postcssPresetPrimer()],
       postcssModulesOptions,
     }),
